Add a CMakeLists.txt file for use by cmake, ctest
This enables out-of-tree builds, and the use of other build tools.

This will not overwrite Makefile if you are building in the
source tree.  To use makefiles generated from CMake, you must
build out of the source tree.

Usage: If SPIRV-Cross source is in $SPIRV_CROSS_ROOT, then to
build with ninja in a different directory:

   cmake -G Ninja $SPIRV_CROSS_ROOT
   ninja

   # Run tests, if enabled
   ctest

Tests will be enabled if you have python3.