MSL Argument Buffers: preserve all resources
In MSL, the unused contents of an argument buffer are important if that argument buffer is to be used across multiple entry points; unused members still contribute to the layout of used members. To accomodate use cases where an argument buffer is reused across multiple entry points (e.g. bound to the vertex and fragment stages at the same time), don't strip unused resources when argument buffers are enabled.

This is a tentative change for review. If the basic idea looks good then I can go through and update the tests. Otherwise, if this needs to be behind a flag, let me know and I will alter the patch accordingly.